I added the frontend components for purchasing the tickets. I added the following:
Purchase-Screen.tsx: This page will show all of the payment details before confirming payment. It uses an example event object to fill in the following details such as event name, etc. The numbers are static for now but we could add pricing to the event object to make things more dynamic.

-PurchasePopup.tsx: This is the component that pops up on the screen when confirming purchase. This component can be reused for other things.

ticket-screen.tsx: This page shows the ticket for scanning into the event. Still static values, but I setup an object for this page called Ticket that will make it easier to integrate it with the backend.

Primary Button: I added extra props to make this button more customizable. You can remove the icon, change bg color, and change the icon from the materialcommunityicons